some filters moved to generator-filters
