+++ template/src/api/routes/$$channel$$.js
@@ -1,6 +1,6 @@
 const Router = require('hermesjs/lib/router');
 const router = new Router();
-const {{ channelName | camelCase }}Handler = require('../handlers/{{ channelName | filenamify }}');
+const {{ channelName | camelCase }}Handler = require('../handlers/{{ channelName | convertToFilename }}');
 module.exports = router;
 {% if channel.hasPublish() %}
   {%- if channel.publish().summary() %}
